Added a /casino command to see stats
