+++ src/main/java/com/ambrosia/add/Ambrosia.java
@@ -6,9 +6,12 @@ import com.ambrosia.add.api.AmbrosiaAPI;
 import com.ambrosia.add.database.AmbrosiaDatabase;
 import com.ambrosia.add.discord.DiscordModule;
 import java.util.List;
+import net.dv8tion.jda.api.interactions.components.buttons.Button;
+import org.jetbrains.annotations.NotNull;
 
 public class Ambrosia extends ApplePlugin {
 
+    public static final String DISCORD_INVITE_LINK = "https://discord.gg/tEAy2dGXWF";
     private static Ambrosia instance;
 
     public static void main(String[] args) {
@@ -21,6 +24,10 @@ public class Ambrosia extends ApplePlugin {
 
     public static Ambrosia get() {
         return instance;
+    }
+    @NotNull
+    public static Button inviteButton() {
+        return Button.link(Ambrosia.DISCORD_INVITE_LINK, "Ambrosia Discord Server");
     }
 
     @Override
