+++ src/main/java/com/ambrosia/add/database/AmbrosiaDatabase.java
@@ -3,6 +3,7 @@ package com.ambrosia.add.database;
 import apple.lib.modules.AppleModule;
 import apple.lib.modules.configs.factory.AppleConfigLike;
 import com.ambrosia.add.Ambrosia;
+import com.ambrosia.add.database.casino.GameResultAggregate;
 import com.ambrosia.add.database.client.ClientDiscordDetails;
 import com.ambrosia.add.database.client.ClientEntity;
 import com.ambrosia.add.database.client.ClientMinecraftDetails;
@@ -46,9 +47,13 @@ public class AmbrosiaDatabase extends AppleModule {
         dbConfig.setDataSourceConfig(dataSourceConfig);
         dbConfig.setDdlGenerate(true);
 //        dbConfig.setDdlRun(true);
-        dbConfig.addAll(
-            List.of(TransactionEntity.class, GameResultEntity.class, ClientEntity.class, ClientMinecraftDetails.class,
-                ClientDiscordDetails.class));
+
+        // tables
+        dbConfig.addAll(List.of(TransactionEntity.class, GameResultEntity.class, ClientEntity.class));
+        // embedded
+        dbConfig.addAll(List.of(ClientMinecraftDetails.class, ClientDiscordDetails.class));
+        // aggregates
+        dbConfig.addAll(List.of(GameResultAggregate.class));
         return dbConfig;
     }
