+++ src/main/java/com/ambrosia/add/database/casino/game/CasinoGameCountByConclusion.java
@@ -1,0 +1,5 @@
+package com.ambrosia.add.database.casino.game;
+
+public record CasinoGameCountByConclusion(long count, long deltaWinnings) {
+
+}
