+++ src/main/java/com/ambrosia/add/database/casino/game/CasinoGamesCount.java
@@ -1,0 +1,7 @@
+package com.ambrosia.add.database.casino.game;
+
+import java.util.Map;
+
+public record CasinoGamesCount(Map<String, CasinoGamesCountByName> name) {
+
+}
