+++ src/main/java/com/ambrosia/add/database/casino/profits/CasinoTotalProfitsByOperation.java
@@ -1,0 +1,7 @@
+package com.ambrosia.add.database.casino.profits;
+
+import com.ambrosia.add.database.operation.TransactionType;
+
+public record CasinoTotalProfitsByOperation(TransactionType operationType, long sumAmount) {
+
+}
