+++ src/main/java/com/ambrosia/add/discord/DiscordModule.java
@@ -2,13 +2,15 @@ package com.ambrosia.add.discord;
 
 import apple.lib.modules.AppleModule;
 import apple.lib.modules.configs.factory.AppleConfigLike;
-import com.ambrosia.add.discord.delete.CommandDelete;
+import com.ambrosia.add.discord.commands.CommandHelp;
+import com.ambrosia.add.discord.commands.casino.CommandCasino;
+import com.ambrosia.add.discord.commands.delete.CommandDelete;
+import com.ambrosia.add.discord.commands.operation.CommandCash;
+import com.ambrosia.add.discord.commands.profile.CommandLink;
+import com.ambrosia.add.discord.commands.profile.CreateProfileCommand;
+import com.ambrosia.add.discord.commands.profile.ProfileCommand;
+import com.ambrosia.add.discord.commands.profile.ViewProfileCommand;
 import com.ambrosia.add.discord.log.DiscordLog;
-import com.ambrosia.add.discord.operation.CommandCash;
-import com.ambrosia.add.discord.profile.CommandLink;
-import com.ambrosia.add.discord.profile.CreateProfileCommand;
-import com.ambrosia.add.discord.profile.ProfileCommand;
-import com.ambrosia.add.discord.profile.ViewProfileCommand;
 import com.github.AndrewAlbizati.Blackjack;
 import discord.util.dcf.DCF;
 import discord.util.dcf.DCFCommandManager;
@@ -49,7 +51,7 @@ public class DiscordModule extends AppleModule {
         DCFCommandManager dcfCommands = dcf.commands();
         dcfCommands.addCommand(new CreateProfileCommand(), new ProfileCommand(), new ViewProfileCommand(), new CommandLink());
         dcfCommands.addCommand(new CommandCash());
-        dcfCommands.addCommand(new CommandDelete());
+        dcfCommands.addCommand(new CommandDelete(), new CommandCasino());
         dcfCommands.addCommand(new CommandHelp());
         new DiscordLog(dcf);
     }
